updated code: the new version had a bug in the Kadanoff-Baym equatuion solution introduced with the discretization of the functions arguments in the vide_rkX routines. I restored the older version with function as arguments and checked that the code gives exactly the same results in the non-interacting limit as the test drivers testGFviderk2 and neqGZ codes.
update: testing the U!=0 case.
G.Mazza suggested that the <-component should be calculated only in the
upper triangular regime, which decreases the calculation time.

The possible large error along the diagonal, related to the sum of diagonal in k,
should come from the double errors: one on the initial conditions and one over the dynamics.
The former are in fact obtained from the dynamics along the first line (t`=0).
Can we do better?