debug: keep testing this new version of the code. I fixed some issues with the evaluation of Sigma function (IPT) from G0. The gtr component was wrongly evaluated from less and ret component which generated a bad Sigma function, compromising the entire calculation.
Other fixing in the code in BATH, KADANOFF-BAYM, IPT.

Rational @260913:
I am testing the iteration with and without the interactino and bath self-energy.
Somehow (and here is the bug I think), in presence of BOTH the self-energies
the system does not conserve the number of particles. In passing I should notice that
this bug remind very much to the gapped/pseud-gapped bath issue of non-conserving density.
If U=0/V!=0 OR V=0/U!=0 the density is conserved. Note that in the first case we are
including the effect of coupling quench with the thermostat.