NEW VERSION OF THE CODE. This version is based on causal evoution of the local function. All the equations are reduced to integral or integro-differential equations.
A large class/module for the contour GF is written, including mixed gf.

The code is now very short and all included in almost one file in the driver dir.

In this version it keeps using IPT solver. As a starting guess we use G0 (the Weiss Field)
but this requires the user to write a specific junction between input and the solver, basically
reshaping the dmft-loop with respect to my usual implmentation (SC--->IS), now is IS--->SC.

Checking the stability of the code it seems like that the code performs better at larger U if the
matsubara/imaginary time integral are performed with a larger basis. Namely because the self-energy
is very sharp at the boundaries (0,beta). I will implement a version of the code using PQ-basis.

Another driver is in the middle of development to start from a given self-energy Matsubara.
This requires to Fourier transforms to imaginary time, which is only possible if nasty tail C_0 + C_1/iw
are removed and transformed alone.
Checking on this.