CODE WORKING The new version of the code using causal propagation at 2nd order and VIE/VIDE solution has been thoroughly tested against the available data for the quench. [This version of the code is based on the algorithm provided by M.Eckstein, P.Werner, N.Tsuji and Co., plus some minor improvements]
The main driver: +neqdmft_bethe_dos was shown to give identical results
as its related code neqdmft_bethe performing algebraic self-consistency on
the bethe lattice. The results are identical to published data.

Tests for the quench in the 2D square lattice are ok.
Test for the electric field effects on the 2D square lattice are going.