code working up to now. bethe and bethe_dos gives same results as for the old data available of the test dir.
there is something wrong anyway:
the comparison of the results plotted in the file cfr_BetheDensityVSBetheAlgebraic_Quench.agr
shows that analytic and density self-consistency gave same results, but the data in the test
results, including the old ones obtained with old version of the code, are different one against
another.

something fishy here.
though it may be that: there was an error in the plot and the data normally look slightly different
because the integration over DOS is somehow different than the algebraic self-consistency.

Last check is to allows FFTGF to work as in the old times and try to compile the first version of the code.