reshape almost finished.
tests for 2dsquare_field is excellent.

few things todo:
I think it would be a good idea to wrap-up a special FFT for this code, starting
from the ones available in FFTGF. Now the code (in AUX_FUNX) uses the older
version of the transforms automatically selected as coefficients C is not passed
as an argument, this means the tails are treated in the old way for the iw-->tau
transform, while not treated at all in the inverse transform tau-->iw.
I should test what happens and how it improves is tails are used.
In any case this version performs transform using longer arrays (Lfreq) and
extract the results on the short (Ntau) array using local version of extract_gtau.