UPDATE: we removed the calculation of the Kernel in convolute_kb_contour_gf along the Horizontal line. The only component involved in the lesser function.
The knowledge of a Kernel K = A*B along this line (i,N), i=1,...,N-1 is irrelevant with
respect to all of the other procedure involved in the solution of neq-DMFT (VIDE,VIE,other...).

This fact can be exploited to reduce the memory usage by solving the equation only for the
lower triangle in t,t` plane (t>t`) and allocating Keldysh functions only in this subspace.