UPDATE:
We terminated the writing of the driver program for the superconducting
non-equilibrim dynamics, attractive Hubbard model.
The code has to be entirely tested.

During the process we introduced numerous changes to the original code (NORMAL case).
All the changes will thoroughly reviewed in the final commit of this devel branch,
and in any case before the merge with master branch.

In this update we introduced the possibility to add and convolute
multiple functions recursively in the form:

C = A1 + A2 + A3 + ... + AN
C = A! * A1 * A3 + ... * AN

which came particularly handy in the writing of this code.

TODO:
The convolution now DO NOT get the horizontal line of the perimeter
in the square domain {t,t'}, this is possible as soon as convolutions
are performed always from the left to right.
In general it is possible to reduce the code to evolve only the vertical
part of the perimeter (N,j=1,...,N) also in the VIDE, VIE and all that.
Even more in general it is possible to store only the lower triangle part
of the functions (saving approx. 50% of the memory), provided
all the integral which involved terms in the upper triangular part are
changed. This algorhtm was successfully tested  by G.Mazza for the
case of the non-equilibrium dynamics in the correlated heterostructure within
Gutzwiller approx.