update: we updated the code with the latest development from the master branch: - introduced sum/del to sum or reset (delete) functions along the perimeter of the contour.
The code compile and is going to be tested.
There is a small issue to solve related to the way to handle the HFB part
of the self-energy, especially in the initialization part of the code.

Following a suggestion of F.Peronaci, we decided to separate completely the
"regular" (Sigma_reg = Sigma-Sigma_HFB) (in this version only the 2nd order
term) from the HFB part. The two terms will be handled differently throughout
the entire code.

AA: I will now test this modification for the normal case, where this is
totally redundant for the half-filling case. But will help clarifying the
details of the things to change without messing up with the code.
Then I will come back and implement this change into the devel_superc code.