update:
the code has been updated with a separation of the self-energy function
in two parts:
- Sigma_Hartree_Fock(_Bogoliubov for SC channel) which is local in time, i.e.
proportional to a delta function on the contour
- Sigma_regular which is the regular part of the self-energy on the contour.

The Sigma_HF is included in the Hamiltonian when solving for a VIDE.
The regulart part is used in the Kernel.

The regular part is also used to evaluate the HF-corrected Weiss Field.

Although this modification was irrelevant for the half-filling case here,
because the Sigma_HF is null in this case, this becomes very handy in the
superconducting channel where the anomalous part of Sigma_HF is always
non-zero.