update:
we updated the code with the latest development from
test_split_sigma.
Namely we divided the self-energy in two parts Sigma_hartree-fock-bogoliubov
and a regular part (2nd order IPT).
All the other features have been imported from test_split_sigma branch.
Some bugs have been fixed in the process.

Code compiled and 1st test for the non-interacting and non-broken-symmetry done.
More tests to come for complete debug.

update:
further tests show that there is a problem in the initialization of the code.
To fix the issue, it is realized that a better strategy to init the code would
be to read the self-energy(ies) directly in imginary time tau\in[0,beta].
- all equilibrium solvers can in principle provide this information, if not
directly at least with a Fourier Transform.
- there will be no need to keep matsubara freq. functions other than to
reconstruct the Gloc and G0 for the equilibrium, which can be done easily.