update of the code:
the init part of the code has been deeply updated.
After having separated the self-energy in two parts (HF and Regular part),
we realized that the two can happily live separated in the time (imaginary
and real) domain.
This is also true in the Matsubara freq. formalism, but there it is
more handy to keep the two united though this introduced a slight
asymmetry.

So: in this new version we directly read Sigma(tau) (the regular part) as
it comes out from a honest equilibrium DMFT calculation. It is soon
realized that every solver can provide this function.
The HF part is generated starting from the knowledge of the static
observales (in this case just the density).
The read self-energy goes into the regular part, it is continued
to the Keldysh contour and filled through the dynamics.
At each step the HF component is also evaluated.

This approach allows to get rid of all the delta functions appearing
in the equations. The only exceptions come from the cases:
A * C.delta
that must be handled directly using a special extension of the convolution.

The code has been tested and gives same results as before.

update: the initial double occupancy is evaluated using the Matsubara freq.
components, which provide a better calculation (Lf >> Ntau).