UPDATE: this is an updated version with respect to the latest. Changes have been imported in the init process. Self-energy functions are read from file directly in the imginary time formalism. Initialization respects the paradigm of keeping separated the HF and Regular components of the self-energies.
The t=t'=0 case has been tested and many bugs removed.
THe dynamics seems to work but tests have been performed with Sigma=0.

The HF part is not included in all steps in the driver program now, this
needs a small rewriting to take care of the separation into HF and Reg part.

In practice the code always requires to use Sigma==Sigma_reg as a kernel for
the dynamics, while the HF part is included in the Hamiltonian term.
Nevertheless, some the HF appears now in some convolutions in the form:

A = (Sigma_HF.delta_C + Sigma) * B

My approach would be to get rid of the delta functions everywhere by
performing the multiplication separately, rather than performing the sum
in the parentheses first.
But this approach is very lenghty and prone to errors.

Can we find an optimal way to solve this kind of convolutions?
Because of the structure of the derived types (allocatable) it seems
not plausible to overload the + and * operations to sum and convolutions so
to be able to perform the algebraic operation as usual.