big update:
This version feature a big update and reshape of the code.
The NEQ_CONTOUR_GF.f90 module, which is the core of the code, has been
splitted in many different files, each containing dedicated procedures
for a specific task, e.g. neq_contour_gf_convolute.f90 contains
routines that performs various type of convolutions.

In particular a kb_contour_sigma type has been introduced to contain
self-energy functions, which includes regular, i.e. a normal kb_contour_gf
type, plus a singular part, i.e. a term proprtional to a contour delta
function (local in time).
Quite all the routines have been extended to this new type, in particular
the convolutions and vide/vie solution. In this way it was straightforward
to automatically perform product of the form:
C= (Sigma_HF.delta_C(t,t') + Sigma(t,t'))*B

as well as more simple products

C = A*B

Similarly, the solution of vide/vie with a Self-energy kernel K is automated,
so that the HF, singular part goes straight with the linear term of the form h*f,
while the regular part actually define the kernel K in K*f.

------->     WARNING      <---------
This approach requires to modify a bit the drivers to comply with the new standard.
While this is a super easy step, it allows to extend the codes to more complicated
situations, just like the away from half-filling case or different way to treat the
HF, singular, term.

Test for the BETHE-DOS case has been performed with 100% success.

The ATTRACTIVE case has been adapted to the new code and compiled.
Test are in order with Sigma_regular == 0, i.e. solve the static mean-field
dynamics first.