v2.4.0 UPDATE
This is a newer version of the code, developed with Angelo.
We introduced slight modifications to the structure of the code,
e.g. we modified the equilibrium starting point by introducing EQ_IPT
module (taken from DMFT_IPT code).
We updated and tested the driver neqipt_2dsquare_field solving first the
equilibrium problem and the non-eq dynamics.

Tests seem ok.
Next step would be to include the superconductive formalism based on
Nambu structure of the GF,Self.